% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_mod11}
\alias{check_mod11}
\title{Check the modulo 11 rules for CPR numbers}
\usage{
check_mod11(cpr)
}
\arguments{
\item{cpr}{A vector of CPR number strings}
}
\value{
A logical vector with values TRUE (CPR number ok), FALSE (CPR number not ok), or NA (cannot check)
}
\description{
Check the modulo 11 rules for CPR numbers
}
