% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{clean_cprstring}
\alias{clean_cprstring}
\title{Check the modulo 11 rules for CPR numbers}
\usage{
clean_cprstring(cpr)
}
\arguments{
\item{cpr}{A vector of potential CPR number strings}
}
\value{
A vector of cleaned CPR number strings
}
\description{
Check the modulo 11 rules for CPR numbers
}
\examples{
cpr <- c("1010101234", "11111111", "101010-1234", "Cprnr : 101010-1234")
clean_cprstring(cpr)
}
