% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dob.R
\name{date_of_birth}
\alias{date_of_birth}
\title{Returns the date-of-birth from a CPR number}
\usage{
date_of_birth(cpr)
}
\arguments{
\item{cpr}{A vector of CPR numbers strings}
}
\value{
A Date object with the date-of-births in YYYY-MM-DD format or NA if not a proper date
}
\description{
Returns the date-of-birth from a CPR number
}
\examples{
date_of_birth("2310450637")  # This is testing out a randomly generated number
}
