% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{gender}
\alias{gender}
\title{Returns the gender from a CPR number}
\usage{
gender(cpr)
}
\arguments{
\item{cpr}{A vector of CPR numbers strings}
}
\value{
A vector with gender as integer (0 = female, 1 = male)
}
\description{
Gender is encoded in the last digit. If it is odd then it is a male, and if it is even then it is female.
}
\examples{
gender("2310450637")  # This is testing out a randomly generated number

}
