% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_cpr}
\alias{is_cpr}
\title{Check whether the CPR number is valid}
\usage{
is_cpr(cpr)
}
\arguments{
\item{cpr}{A vector of cpr number strings}
}
\value{
A logical vector with values TRUE (cpr number ok), FALSE (cpr number not ok), or NA (Not a proper date)
}
\description{
Check whether the CPR number is valid
}
