% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{valid_cpr}
\alias{valid_cpr}
\title{Check the modulo 11 rules for CPR numbers}
\usage{
valid_cpr(cpr)
}
\arguments{
\item{cpr}{A vector of cpr number strings}
}
\value{
A logical vector with values TRUE (cpr number ok), FALSE (cpr number not ok), or NA (Not a proper date)
}
\description{
Check the modulo 11 rules for CPR numbers
}
